pmacc: remove folder `nvidia`
- move files from folder `nvidia'into `device`, `kernel` and
`math/operation`
- update includes and namespaces
