+++ include/pmacc/math/operation/Max.hpp
@@ -22,30 +22,32 @@
 #pragma once
 
 #include "pmacc/mpi/GetMPI_Op.hpp"
+#include "pmacc/algorithms/math.hpp"
+
 #include "pmacc/types.hpp"
 
 namespace pmacc
 {
-    namespace nvidia
+    namespace math
     {
-        namespace functors
+        namespace operation
         {
-            struct Add
+            struct Max
             {
                 template<typename Dst, typename Src>
-                HDINLINE void operator()(Dst& dst, const Src& src) const
+                DINLINE void operator()(Dst& dst, const Src& src) const
                 {
-                    dst += src;
+                    dst = math::max(dst, src);
                 }
 
                 template<typename Dst, typename Src, typename T_Acc>
-                HDINLINE void operator()(const T_Acc&, Dst& dst, const Src& src) const
+                DINLINE void operator()(const T_Acc&, Dst& dst, const Src& src) const
                 {
-                    dst += src;
+                    dst = math::max(dst, src);
                 }
             };
-        } // namespace functors
-    } // namespace nvidia
+        } // namespace operation
+    } // namespace math
 } // namespace pmacc
 
 namespace pmacc
@@ -53,9 +55,9 @@ namespace pmacc
     namespace mpi
     {
         template<>
-        HINLINE MPI_Op getMPI_Op<pmacc::nvidia::functors::Add>()
+        HINLINE MPI_Op getMPI_Op<pmacc::math::operation::Max>()
         {
-            return MPI_SUM;
+            return MPI_MAX;
         }
     } // namespace mpi
 } // namespace pmacc
