+++ include/pmacc/math/operation/Mul.hpp
@@ -1,4 +1,4 @@
-/* Copyright 2013-2021 Heiko Burau, Rene Widera, Benjamin Worpitz
+/* Copyright 2014-2021 Axel Huebl
  *
  * This file is part of PMacc.
  *
@@ -21,33 +21,31 @@
 
 #pragma once
 
-#include "pmacc/mpi/GetMPI_Op.hpp"
-#include "pmacc/algorithms/math.hpp"
-
 #include "pmacc/types.hpp"
+#include "pmacc/mpi/GetMPI_Op.hpp"
 
 namespace pmacc
 {
-    namespace nvidia
+    namespace math
     {
-        namespace functors
+        namespace operation
         {
-            struct Max
+            struct Mul
             {
                 template<typename Dst, typename Src>
-                DINLINE void operator()(Dst& dst, const Src& src) const
+                HDINLINE void operator()(Dst& dst, const Src& src) const
                 {
-                    dst = math::max(dst, src);
+                    dst *= src;
                 }
 
                 template<typename Dst, typename Src, typename T_Acc>
-                DINLINE void operator()(const T_Acc&, Dst& dst, const Src& src) const
+                HDINLINE void operator()(const T_Acc&, Dst& dst, const Src& src) const
                 {
-                    dst = math::max(dst, src);
+                    dst *= src;
                 }
             };
-        } // namespace functors
-    } // namespace nvidia
+        } // namespace operation
+    } // namespace math
 } // namespace pmacc
 
 namespace pmacc
@@ -55,9 +53,9 @@ namespace pmacc
     namespace mpi
     {
         template<>
-        HINLINE MPI_Op getMPI_Op<pmacc::nvidia::functors::Max>()
+        HINLINE MPI_Op getMPI_Op<pmacc::math::operation::Mul>()
         {
-            return MPI_MAX;
+            return MPI_PROD;
         }
     } // namespace mpi
 } // namespace pmacc
