+++ include/pmacc/particles/operations/CountParticles.hpp
@@ -28,7 +28,7 @@
 
 #include "pmacc/particles/particleFilter/FilterFactory.hpp"
 #include "pmacc/particles/particleFilter/PositionFilter.hpp"
-#include "pmacc/nvidia/atomic.hpp"
+#include "pmacc/kernel/atomic.hpp"
 #include "pmacc/memory/shared/Allocate.hpp"
 #include "pmacc/traits/GetNumWorkers.hpp"
 #include "pmacc/mappings/threads/ForEachIdx.hpp"
@@ -118,7 +118,7 @@ namespace pmacc
                         {
                             auto parSrc = (frame[linearIdx]);
                             if(accParFilter(acc, parSrc))
-                                nvidia::atomicAllInc(acc, &counter, ::alpaka::hierarchy::Threads{});
+                                kernel::atomicAllInc(acc, &counter, ::alpaka::hierarchy::Threads{});
                         }
                     }
                 });
